#1352 removed scala shell interface but not implementation
